proposition for lattices
